+++ src/controllers/api/SemesterEducationFormApiController.php
@@ -2,16 +2,22 @@
 
 namespace App\Controllers;
 
+require_once __DIR__ . '/../BaseController.php';
+require_once __DIR__ . '/../../services/WPService.php';
 require_once __DIR__ . '/../../services/SemesterEducationFormService.php';
 
+use App\Controllers\BaseController;
+use App\Services\WPService;
 use App\Services\SemesterEducationFormService;
 
-class SemesterEducationFormApiController
+class SemesterEducationFormApiController extends BaseController
 {
 	protected SemesterEducationFormService $semesterEducationFormService;
+	protected WPService $wpService;
 
 	function __construct()
 	{
+		$this->wpService = new WPService();
 		$this->semesterEducationFormService = new SemesterEducationFormService();
 	}
 
@@ -25,9 +31,14 @@ class SemesterEducationFormApiController
 		$educationalDisciplineSemesterId = intval($data['educationalDisciplineSemesterId']);
 		$educationalFormId = intval($data['educationalFormId']);
 
-		$this->semesterEducationFormService->createSemesterEducationForm($educationalDisciplineSemesterId, $educationalFormId);
+		$wpCreatorId = $this->wpService->getWPCreatorIdBySemesterId($educationalDisciplineSemesterId);
+		$ifCurrentUserHasAccessToWP = $this->checkIfCurrentUserHasAccessToWP($wpCreatorId);
 
-		echo json_encode(['status' => 'success', 'message' => 'educationalDisciplineSemesterEducationForm was created']);
+		if ($ifCurrentUserHasAccessToWP) {
+			$this->semesterEducationFormService->createSemesterEducationForm($educationalDisciplineSemesterId, $educationalFormId);
+
+			echo json_encode(['status' => 'success', 'message' => 'educationalDisciplineSemesterEducationForm was created']);
+		}
 	}
 
 	public function deleteSemesterEducationForm()
@@ -37,6 +48,11 @@ class SemesterEducationFormApiController
 		$educationalDisciplineSemesterId = $_GET['educationalDisciplineSemesterId'];
 		$educationalFormId = $_GET['educationalFormId'];
 
-		$this->semesterEducationFormService->deleteSemesterEducationForm($educationalDisciplineSemesterId, $educationalFormId);
+		$wpCreatorId = $this->wpService->getWPCreatorIdBySemesterId($educationalDisciplineSemesterId);
+		$ifCurrentUserHasAccessToWP = $this->checkIfCurrentUserHasAccessToWP($wpCreatorId);
+
+		if ($ifCurrentUserHasAccessToWP) {
+			$this->semesterEducationFormService->deleteSemesterEducationForm($educationalDisciplineSemesterId, $educationalFormId);
+		}
 	}
 }
